+++ src/main/java/com/codifica/chatbot/infrastructure/persistence/chat/ChatJpaRepository.java
@@ -2,6 +2,6 @@ package com.codifica.chatbot.infrastructure.persistence.chat;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatJpaRepository extends JpaRepository<ChatEntity, Integer> {
+public interface ChatJpaRepository extends JpaRepository<ChatEntity, Long> {
 
 }
